Added styles class this needs to be used for all object Wesley i added a few of this code in the constructor see how im using it.
The purpose is basically to have 1 main style and every button and frame and panel etc inherit from this class.

Added the main form that you all been waiting for so anxiously.it goes to the other forms but i will modify this as time goes on.

Do not copy and paste colors and formatting, use the style class to do this.

COOL!! enjoy.